mplayer.py: correctly handle video expanding.
With "-vf expand", the old subtitle renderer (a.k.a. "-noass") works
just fine; while the libass renderer ("-ass") fails to flush the
display, causing never-disappeard and overlapped subtitles.

With the "-ass-use-margins" video expander, the libass renderer does
flush the display. But it has its own issue: the subtitle fonts are
rendered at a wrong aspect - they are instead of a square!